Reorganize operators to their own file
